requirements.txt
tensorflow==1.13.1
keras==2.2.4
opencv-python==4.1.1.26
Pillow==6.1.0
matplotlib==3.1.1
configparser
imgaug
